+++ assnake/config_internal.ini
@@ -1,3 +1,3 @@
 [GENERAL]
-config_loc = /home/fedor/assnake/config.yaml
+config_loc = None
