+++ assnake/snake/config_template.yml
@@ -39,42 +39,3 @@ MetaPhlAn2:
 
     # mpa_v20_m200: '/data5/bio/databases/MetaPhlAn2/mpa_v20_m200/mpa_v20_m200.pkl'
     # BOWTIE2DB: '/data5/bio/databases/MetaPhlAn2/mpa_v20_m200'
-# METAPHLAN2: '/data3/bio/biouser/tools/metaphlan2/metaphlan2.py'
-# MPA_PKL: '/data3/bio/biouser/tools/metaphlan2/db_v20/mpa_v20_m200.pkl'
-# BOWTIE2DB: '/data3/bio/biouser/tools/metaphlan2/db_v20/mpa_v20_m200'
-# prodigal: '/data5/bio/runs-fedorov/tools/prodigal_2.6.3/prodigal'
-# hmmer: 
-#     folder: '/data5/bio/runs-fedorov/tools/hmmer/hmmer-3.1b2-linux-intel-x86_64/binaries/'
-#     pfam: '/data5/bio/runs-fedorov/tools/hmmer/db/Pfam-A.hmm'
-
-# gatk.jar : '/data7a/bio/BacGenomeSoft/GenomeAnalysisTK.jar'
-# picard.jar : '/data7a/bio/BacGenomeSoft/picard.jar'
-
-
-# trimmomatic:
-#     bin: "/data6/bio/TFM/soft/Trimmomatic-0.36/trimmomatic-0.36.jar"
-# fastqc.bin: '/data6/bio/TFM/soft/FastQC/fastqc'
-# bbduk: '/data6/bio/TFM/soft/bioinf/mapping/bbmap-38.34/bbduk.sh'
-# bbmap.bin: '/home/fedorov/tools/bbmap/bbmap.sh'
-# bowtie2.bin: /srv/common/bin/bowtie2
-# bowtie2-build.bin: /srv/common/bin/bowtie2-build
-# bowtie2.additional.params: -k 1 
-# bowtie2.threads: '20'
-
-# SOFTWARE
-# java.bin: '/srv/common/opt/jdk1.8.0_91/bin/java'
-# megahit.bin: '/data6/bio/TFM/soft/megahit/megahit'
-# bwa.bin : '/srv/common/bin/bwa'
-# samtools.bin: '/srv/common/bin/samtools'
-# anvio.bin : '/data4/bio/fedorov/miniconda3/envs/anvio5/bin/'
-
-
-# FindFungi:
-#     v0.23.3:
-#         consensus: '/data6/bio/TFM/pipeline/bin/scripts/FindFungi/Kraken32-to-Consensus.py'
-#         taxids: '/data6/bio/TFM/pipeline/bin/scripts/FindFungi/Fungal-Taxids.txt'
-# min_contig_size : 1000
-
-# TEST DATASET
-# test_df_prefix: '/data6/bio/TFM/datasets'
-# test_df_name: 'mg_test'
