+++ assnake/snake/snake_base.py
@@ -1,9 +1,8 @@
 import glob, os, pkg_resources, time
 import assnake.utils
-
-
+print('dsffdsf')
 wc_config = assnake.utils.load_wc_config()
-
+# print(wc_config)
 start = time.time()
 
 # Discover plugins
@@ -11,6 +10,8 @@ discovered_plugins = {
     entry_point.name: entry_point.load()
     for entry_point in pkg_resources.iter_entry_points('assnake.plugins')
 }
+
+
 
 # We need to update wc_config first
 for module_name, module_class in discovered_plugins.items():
@@ -22,8 +23,9 @@ for module_name, module_class in discovered_plugins.items():
 # and now include all the stuff
 for module_name, module_class in discovered_plugins.items():
     for snakefile in module_class.snakefiles:
-        include: os.path.join(module_class.install_dir, snakefile)
-
+        print(os.path.normpath(os.path.join(module_class.install_dir, snakefile)))
+        include: os.path.normpath(os.path.join(module_class.install_dir, snakefile))
+# include: '/mnt/c/Users/fedor/Documents/GitHub/assnake-core-assembly/assnake_core_assembly/megahit/megahit.py'
 
 end = time.time()
 if config['debug']:
